refactor: ディレクトリ構造をさらに簡略化
- application/use-cases/articles/ → application/articles/
- application/use-cases/bookmarks/ → application/bookmarks/
- domain/articles/Article.ts → infrastructure/domain/Article.ts
- domain/bookmarks/Bookmark.ts → infrastructure/domain/Bookmark.ts
- 全てのimportパスを新しいディレクトリ構造に更新
