+++ graphql/src/application/articles/FetchArticlesUseCase.ts
@@ -1,0 +1,8 @@
+import * as articleRepository from "../../infrastructure/external/ArticleRepositoryImpl";
+import type { Article } from "../../infrastructure/domain/Article";
+
+export const fetchArticlesUseCase = async (
+  page: number,
+): Promise<Article[]> => {
+  return articleRepository.fetchArticles(page);
+};
