--- graphql/src/application/use-cases/articles/FetchArticlesUseCase.ts
@@ -1,8 +1,0 @@
-import type { Article } from "../../../domain/articles/Article";
-import * as articleRepository from "../../../infrastructure/external/ArticleRepositoryImpl";
-
-export const fetchArticlesUseCase = async (
-  page: number,
-): Promise<Article[]> => {
-  return articleRepository.fetchArticles(page);
-};
