--- graphql/src/application/use-cases/bookmarks/FetchBookmarkByIdUseCase.ts
@@ -1,8 +1,0 @@
-import type { Bookmark } from "../../../domain/bookmarks/Bookmark";
-import * as bookmarkRepository from "../../../infrastructure/persistence/BookmarkRepositoryImpl";
-
-export const fetchBookmarkByIdUseCase = async (
-  id: string,
-): Promise<Bookmark | null> => {
-  return bookmarkRepository.findById(id);
-};
