--- graphql/src/application/use-cases/bookmarks/FetchBookmarksUseCase.ts
@@ -1,6 +1,0 @@
-import type { Bookmark } from "../../../domain/bookmarks/Bookmark";
-import * as bookmarkRepository from "../../../infrastructure/persistence/BookmarkRepositoryImpl";
-
-export const fetchBookmarksUseCase = async (): Promise<Bookmark[]> => {
-  return bookmarkRepository.findMany();
-};
