+++ graphql/src/infrastructure/domain/Article.ts
