+++ graphql/src/infrastructure/domain/Bookmark.ts
